backend/requirements.txt
# Surface backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
openai>=1.0.0
httpx>=0.27.0
